blog-service/pyproject.toml
[tool.poetry]
name = "blog-service"
version = "0.1.0"
description = "A Django service responsible for main blog operations"
authors = ["duciito <duciito@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "5.0.1"
django-storages = "1.14.2"
djangorestframework = "3.14.0"
django-cors-headers = "3.6.0"
pyjwt = "2.6.0"
redis = "5.0.1"
celery = "5.3.6"
cryptography = "3.4.8"
psycopg = {extras = ["binary"], version = "^3.1.17"}


[tool.poetry.group.dev.dependencies]
ipdb = "^0.13.13"
ruff = "^0.1.11"

[tool.ruff]
select = [ "E", "F", "UP", "ISC", "I", "W", "B" ]
target-version = "py311"

[tool.pyright]
venvPath = "."
venv = ".venv"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
